댓글 리스트 ajax로 불러오기

+++ src/main/java/com/jafa/domain/ReplyVO.java
@@ -16,8 +16,8 @@ import lombok.ToString;
 @ToString
 @Builder
 public class ReplyVO {
-	private int rno; 
-	private int bno; 
+	private Long rno; 
+	private Long bno; 
 	private String reply;
 	private String id;
 	private String nickname;
